+++ src/main/java/com/nisovin/shopkeepers/config/value/ValueTypeProvider.java
@@ -1,0 +1,18 @@
+package com.nisovin.shopkeepers.config.value;
+
+import java.lang.reflect.Type;
+
+public interface ValueTypeProvider {
+
+	/**
+	 * Gets the {@link ValueType} for the given setting type, if it can provide one.
+	 * <p>
+	 * The returned {@link ValueType} may get cached and used for future requests for the same setting type. The
+	 * provider is therefore only allowed to take the type itself into account, and not any other contextual state.
+	 * 
+	 * @param type
+	 *            the type of the setting's value
+	 * @return the value type, or <code>null</code> if none can be provided
+	 */
+	public ValueType<?> get(Type type);
+}