+++ src/main/java/com/nisovin/shopkeepers/config/value/ValueTypeProviders.java
@@ -1,0 +1,27 @@
+package com.nisovin.shopkeepers.config.value;
+
+import java.lang.reflect.Type;
+import java.util.function.Function;
+
+import com.nisovin.shopkeepers.util.Validate;
+
+public class ValueTypeProviders {
+
+	public static ValueTypeProvider forTypePattern(TypePattern typePattern, Function<Type, ValueType<?>> valueTypeProvider) {
+		Validate.notNull(typePattern, "typePattern is null");
+		Validate.notNull(valueTypeProvider, "valueTypeProvider is null");
+		return new ValueTypeProvider() {
+			@Override
+			public ValueType<?> get(Type type) {
+				if (typePattern.matches(type)) {
+					return valueTypeProvider.apply(type);
+				} else {
+					return null;
+				}
+			}
+		};
+	}
+
+	private ValueTypeProviders() {
+	}
+}