+++ src/main/java/com/nisovin/shopkeepers/config/value/ValueTypeRegistry.java
@@ -1,0 +1,55 @@
+package com.nisovin.shopkeepers.config.value;
+
+import java.lang.reflect.Type;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+import com.nisovin.shopkeepers.util.Validate;
+
+/**
+ * Registry of value types for setting types.
+ */
+public class ValueTypeRegistry {
+
+	private final Map<Type, ValueType<?>> byType = new HashMap<>();
+	// Ordered: The first successful provider is used.
+	private final List<ValueTypeProvider> providers = new ArrayList<>();
+
+	public ValueTypeRegistry() {
+	}
+
+	// Replaces any previously registered ValueType.
+	public <T> void register(Type type, ValueType<? extends T> valueType) {
+		Validate.notNull(type, "type is null");
+		Validate.notNull(valueType, "valueType is null");
+		byType.put(type, valueType);
+	}
+
+	public boolean hasCachedValueType(Type type) {
+		return byType.containsKey(type);
+	}
+
+	public void register(ValueTypeProvider valueTypeProvider) {
+		Validate.notNull(valueTypeProvider, "valueTypeProvider is null");
+		providers.add(valueTypeProvider);
+	}
+
+	@SuppressWarnings("unchecked")
+	public <T> ValueType<T> getValueType(Type type) {
+		ValueType<T> valueType = (ValueType<T>) byType.get(type);
+		if (valueType == null) {
+			// Check providers:
+			for (ValueTypeProvider provider : providers) {
+				valueType = (ValueType<T>) provider.get(type);
+				if (valueType != null) {
+					// Cache result:
+					this.register(type, valueType);
+					break;
+				} // Else: Continue searching.
+			}
+		}
+		return valueType;
+	}
+}